unbundle t-function-matcher plugin
