+++ inlang/source-code/plugins/json/package.json
@@ -30,12 +30,10 @@
 		"@inlang/cli": "*",
 		"@types/flat": "^5.0.2",
 		"@types/lodash.merge": "4.6.7",
-		"@types/parsimmon": "1.10.6",
 		"@vitest/coverage-v8": "^0.33.0",
 		"vitest": "0.34.3",
 		"flat": "^5.0.2",
 		"lodash.merge": "4.6.2",
-		"parsimmon": "1.18.1",
 		"typescript": "5.1.3",
 		"patch-package": "6.4.7"
 	}
