+++ inlang/source-code/plugins/t-function-matcher/src/ideExtension/config.ts
@@ -1,0 +1,31 @@
+import { parse } from "./messageReferenceMatchers.js"
+import type { CustomApiInlangIdeExtension, Plugin } from "@inlang/plugin"
+
+export const ideExtensionConfig = (): ReturnType<Exclude<Plugin["addCustomApi"], undefined>> => ({
+	"app.inlang.ideExtension": {
+		messageReferenceMatchers: [
+			async (args: { documentText: string }) => {
+				return parse(args.documentText)
+			},
+		],
+		extractMessageOptions: [
+			{
+				callback: (args: { messageId: string }) => `{t("${args.messageId}")}`,
+			},
+		],
+		documentSelectors: [
+			{
+				language: "typescriptreact",
+			},
+			{
+				language: "javascript",
+			},
+			{
+				language: "typescript",
+			},
+			{
+				language: "svelte",
+			},
+		],
+	} satisfies CustomApiInlangIdeExtension,
+})
