+++ inlang/source-code/plugins/t-function-matcher/src/index.ts
@@ -1,0 +1,2 @@
+import { plugin } from "./plugin.js"
+export default plugin
