+++ inlang/source-code/plugins/t-function-matcher/src/plugin.ts
@@ -1,0 +1,15 @@
+import { Plugin } from "@inlang/plugin"
+import { displayName, description } from "../marketplace-manifest.json"
+import { PluginSettings } from "./settings.js"
+import { ideExtensionConfig } from "./ideExtension/config.js"
+
+const id = "plugin.inlang.t-function-matcher"
+
+export const plugin: Plugin<{
+	[id]: PluginSettings
+}> = {
+	id,
+	displayName,
+	description,
+	addCustomApi: () => ideExtensionConfig(),
+}
