+++ inlang/source-code/plugins/t-function-matcher/src/settings.ts
@@ -1,0 +1,3 @@
+export type PluginSettings = object
+
+export const PluginSettings = {}
