+++ inlang/source-code/plugins/t-function-matcher/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"include": ["src/**/*"],
+	"compilerOptions": {
+		"resolveJsonModule": true
+	}
+}
